refactor: split Repositories and DataAccess in different namespaces
