+++ Doppler.HtmlEditorApi/DataAccess.DapperProvider/DapperWrapperDbContext.cs
@@ -4,7 +4,7 @@ using System.Data;
 using System.Threading.Tasks;
 using Dapper;
 
-namespace Doppler.HtmlEditorApi.Storage.DapperProvider;
+namespace Doppler.HtmlEditorApi.DataAccess.DapperProvider;
 
 public class DapperWrapperDbContext : IDbContext, IDisposable
 {
