+++ Doppler.HtmlEditorApi/DataAccess/IDbQuery.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Threading.Tasks;
+
+namespace Doppler.HtmlEditorApi.DataAccess;
+
+public interface IDbQuery
+{
+    string GenerateSqlQuery();
+    object GenerateSqlParameters() => this;
+}
+
+public interface IExecutableDbQuery : IDbQuery { }
+
+public interface ICollectionDbQuery<TResult> : IDbQuery { }
+
+public interface ISingleItemDbQuery<TResult> : IDbQuery { }
