+++ Doppler.HtmlEditorApi/Repositories.DopplerDb/Queries/FirstOrDefaultContentWithCampaignStatusDbQuery.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
+using Doppler.HtmlEditorApi.DataAccess;
 
-namespace Doppler.HtmlEditorApi.Storage.DapperProvider.Queries;
+namespace Doppler.HtmlEditorApi.Repositories.DopplerDb.Queries;
 
 public record FirstOrDefaultContentWithCampaignStatusDbQuery(
     int IdCampaign,
