+++ Doppler.HtmlEditorApi/Repositories.DopplerDb/Queries/UpdateCampaignStatusDbQuery.cs
@@ -1,6 +1,6 @@
-using System.Threading.Tasks;
+using Doppler.HtmlEditorApi.DataAccess;
 
-namespace Doppler.HtmlEditorApi.Storage.DapperProvider.Queries;
+namespace Doppler.HtmlEditorApi.Repositories.DopplerDb.Queries;
 
 public record UpdateCampaignStatusDbQuery(
     int setCurrentStep,
